+++ crates/fluvio-connector/src/cli/create.rs
@@ -4,11 +4,12 @@
 //! CLI tree to generate Create a Managed Connector
 //!
 
-use fluvio_controlplane_metadata::connector::ManagedConnectorSpec;
 use structopt::StructOpt;
 use tracing::debug;
 
 use fluvio::Fluvio;
+use fluvio::metadata::topic::{TopicSpec, TopicReplicaParam};
+use fluvio_controlplane_metadata::connector::ManagedConnectorSpec;
 
 use crate::error::ConnectorError;
 use crate::config::ConnectorConfig;
@@ -27,12 +28,17 @@ pub struct CreateManagedConnectorOpt {
 impl CreateManagedConnectorOpt {
     pub async fn process(self, fluvio: &Fluvio) -> Result<(), ConnectorError> {
         let config = ConnectorConfig::from_file(&self.config)?;
-        let spec: ManagedConnectorSpec = config.into();
+        let spec: ManagedConnectorSpec = config.clone().into();
         let name = spec.name.clone();
 
         debug!("creating managed_connector: {}, spec: {:#?}", name, spec);
 
         let admin = fluvio.admin().await;
+        if config.create_topic {
+            let topic_spec = TopicSpec::Computed(TopicReplicaParam::new(1, 1, false));
+            debug!("topic spec: {:?}", topic_spec);
+            admin.create(config.topic, false, topic_spec).await?;
+        }
         admin.create(name.to_string(), false, spec).await?;
 
         Ok(())