+++ crates/fluvio-sc/src/k8/controllers/mod.rs
@@ -1,6 +1,7 @@
 pub mod spg_stateful;
 pub mod spu_service;
 pub mod spu_controller;
+pub mod managed_connector_deployment;
 
 pub use k8_operator::run_k8_operators;
 
@@ -18,9 +19,12 @@ mod k8_operator {
     use crate::k8::objects::statefulset::StatefulsetSpec;
     use crate::k8::objects::spg_service::SpgServiceSpec;
     use crate::k8::objects::spu_k8_config::ScK8Config;
+    use crate::k8::objects::managed_connector_deployment::ManagedConnectorDeploymentSpec;
+
     use crate::k8::controllers::spg_stateful::SpgStatefulSetController;
     use crate::k8::controllers::spu_service::SpuServiceController;
     use crate::k8::controllers::spu_controller::K8SpuController;
+    use crate::k8::controllers::managed_connector_deployment::ManagedConnectorDeploymentController;
 
     pub async fn run_k8_operators(
         namespace: String,
@@ -33,6 +37,9 @@ mod k8_operator {
         let spu_service_ctx: StoreContext<SpuServiceSpec> = StoreContext::new();
         let statefulset_ctx: StoreContext<StatefulsetSpec> = StoreContext::new();
         let spg_service_ctx: StoreContext<SpgServiceSpec> = StoreContext::new();
+        let managed_connector_deployments_ctx: StoreContext<ManagedConnectorDeploymentSpec> =
+            StoreContext::new();
+
         let config_ctx: StoreContext<ScK8Config> = StoreContext::new();
 
         info!("starting k8 cluster operators");
@@ -55,11 +62,17 @@ mod k8_operator {
             spg_service_ctx.clone(),
         );
 
+        K8ClusterStateDispatcher::<_, _>::start(
+            namespace.clone(),
+            k8_client.clone(),
+            managed_connector_deployments_ctx.clone(),
+        );
+
         K8ClusterStateDispatcher::<_, _>::start(namespace.clone(), k8_client, config_ctx.clone());
 
         whitelist!(config, "k8_spg", {
             SpgStatefulSetController::start(
-                namespace,
+                namespace.clone(),
                 config_ctx.clone(),
                 global_ctx.spgs().clone(),
                 statefulset_ctx,
@@ -79,6 +92,13 @@ mod k8_operator {
 
         whitelist!(config, "k8_spu_service", {
             SpuServiceController::start(config_ctx, spu_service_ctx, global_ctx.spgs().clone());
+        });
+        whitelist!(config, "k8_managed_connector_delpoyment", {
+            ManagedConnectorDeploymentController::start(
+                namespace,
+                global_ctx.managed_connectors().clone(),
+                managed_connector_deployments_ctx,
+            );
         });
     }
 }